requirement.txt
streamlit==1.25.0
PyYAML==6.0.1
pandas==2.0.3
numpy==1.25.2
python-dotenv==1.0.0